# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/availability/providers/{providerId}": {
      "get": {
        "operationId": "AvailabilityController_getAvailability",
        "summary": "Generate customer-bookable availability for the configured booking window",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use provider or admin to manage schedules; user to view availability.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProviderAvailability"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "availability"
        ]
      }
    },
...
    "/api/catalog/providers/{id}": {
      "get": {
        "operationId": "CatalogController_findById",
        "summary": "Get one provider from the catalog",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use \"admin\" for full access or \"user\" for read-only GET access.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/CatalogProvider"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "catalog"
        ]
      },
      "patch": {
        "operationId": "CatalogController_update",
        "summary": "Update a provider in the catalog",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use \"admin\" for full access or \"user\" for read-only GET access.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProviderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/CatalogProvider"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "catalog"
        ]
      },
      "delete": {
        "operationId": "CatalogController_delete",
        "summary": "Delete a provider from the catalog",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use \"admin\" for full access or \"user\" for read-only GET access.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/CatalogProvider"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/admin/providers/verification": {
      "get": {
        "operationId": "ProviderVerificationController_findAll",
        "summary": "Get provider verification requests",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use \"admin\" or \"superuser\" to access provider verification.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider verification requests returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProviderVerification"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-provider-verification"
        ]
      },
      "post": {
        "operationId": "ProviderVerificationController_create",
        "summary": "Create provider verification request from provider registration",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use \"admin\" or \"superuser\" to access provider verification.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderVerificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider verification request created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProviderVerification"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-provider-verification"
        ]
      }
    },
    "/api/admin/providers/{id}": {
      "get": {
        "operationId": "ProviderVerificationController_findById",
        "summary": "Get provider verification details",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use \"admin\" or \"superuser\" to access provider verification.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProviderVerification"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-provider-verification"
        ]
      }
    },
    "/api/admin/providers/{id}/documents": {
      "get": {
        "operationId": "ProviderVerificationController_getDocuments",
        "summary": "Get provider verification documents",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use \"admin\" or \"superuser\" to access provider verification.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider documents returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/VerificationDocument"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-provider-verification"
        ]
      }
    },
    "/api/admin/providers/{id}/approve": {
      "patch": {
        "operationId": "ProviderVerificationController_approveProvider",
        "summary": "Approve provider after required documents are approved",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use \"admin\" or \"superuser\" to access provider verification.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider approved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProviderVerification"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-provider-verification"
        ]
      }
    },
    "/api/admin/providers/{id}/reject": {
      "patch": {
        "operationId": "ProviderVerificationController_rejectProvider",
        "summary": "Reject provider verification request",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use \"admin\" or \"superuser\" to access provider verification.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectProviderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider rejected successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProviderVerification"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-provider-verification"
        ]
      }
    },
    "/api/admin/providers/{id}/suspend": {
      "patch": {
        "operationId": "ProviderVerificationController_suspendProvider",
        "summary": "Suspend verified or active provider",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use \"admin\" or \"superuser\" to access provider verification.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider suspended successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProviderVerification"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-provider-verification"
        ]
      }
    },
    "/api/admin/providers/{id}/documents/{documentId}/approve": {
      "patch": {
        "operationId": "ProviderVerificationController_approveDocument",
        "summary": "Approve one provider document",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use \"admin\" or \"superuser\" to access provider verification.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document approved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/VerificationDocument"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-provider-verification"
        ]
      }
    },
    "/api/admin/providers/{id}/documents/{documentId}/reject": {
      "patch": {
        "operationId": "ProviderVerificationController_rejectDocument",
        "summary": "Reject one provider document",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Use \"admin\" or \"superuser\" to access provider verification.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document rejected successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/VerificationDocument"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-provider-verification"
        ]
      }
    },
...
  },
  "info": {
    "title": "ServeEase API",
    "description": "In-memory NestJS backend with CRUD, DTO validation, RBAC, and Swagger.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role": {
        "type": "apiKey",
        "in": "header",
        "name": "role",
        "description": "Role-based access header. Use \"admin\" for full access or \"user\" for read-only GET access."
      }
    },
    "schemas": {
...
      "ProviderAvailability": {
        "type": "object",
        "properties": {
          "providerId": {
            "type": "string",
            "example": "cleanpro-service"
          },
          "bookingWindowDays": {
            "type": "number",
            "example": 7
          },
          "dates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AvailableDate"
            }
          }
        },
        "required": [
          "providerId",
          "bookingWindowDays",
          "dates"
        ]
      },
...
      "CatalogProvider": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cleanpro-service"
          },
          "name": {
            "type": "string",
            "example": "CleanPro Services"
          },
          "category": {
            "type": "string",
            "example": "home-cleaning"
          },
          "subServices": {
            "example": [
              "Kitchen Cleaning",
              "Bathroom Cleaning"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "years": {
            "type": "number",
            "example": 6
          },
          "rating": {
            "type": "number",
            "example": 4.8
          },
          "reviews": {
            "type": "number",
            "example": 245
          },
          "distance": {
            "type": "string",
            "example": "2.5 km"
          },
          "startingPrice": {
            "type": "number",
            "example": 799
          },
          "location": {
            "type": "string",
            "example": "Chennai, Tamil Nadu"
          },
          "jobsDone": {
            "type": "number",
            "example": 520
          },
          "availableToday": {
            "type": "boolean",
            "example": true
          },
          "verified": {
            "type": "boolean",
            "example": true
          },
          "cityId": {
            "type": "number",
            "example": 1
          },
          "image": {
            "type": "string",
            "example": "assets/images/home-cleaning/clean1.jpg"
          },
          "availabilitySlots": {
            "example": [
              "09:00 AM - 11:00 AM"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ownerProviderId": {
            "type": "string",
            "example": "PRO002"
          },
          "ownerProviderEmail": {
            "type": "string",
            "example": "fresh@example.com"
          }
        },
        "required": [
          "id",
          "name",
          "category",
          "subServices",
          "years",
          "rating",
          "reviews",
          "distance",
          "startingPrice",
          "location",
          "jobsDone",
          "availableToday",
          "verified",
          "cityId",
          "image"
        ]
      },
...
      "UpdateProviderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cleanpro-service"
          },
          "name": {
            "type": "string",
            "example": "CleanPro Services"
          },
          "category": {
            "type": "string",
            "example": "home-cleaning"
          },
          "subServices": {
            "example": [
              "Kitchen Cleaning",
              "Bathroom Cleaning"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "years": {
            "type": "number",
            "example": 6
          },
          "rating": {
            "type": "number",
            "example": 4.8
          },
          "reviews": {
            "type": "number",
            "example": 245
          },
          "distance": {
            "type": "string",
            "example": "2.5 km"
          },
          "startingPrice": {
            "type": "number",
            "example": 799
          },
          "location": {
            "type": "string",
            "example": "Chennai, Tamil Nadu"
          },
          "jobsDone": {
            "type": "number",
            "example": 520
          },
          "availableToday": {
            "type": "boolean",
            "example": true
          },
          "verified": {
            "type": "boolean",
            "example": true
          },
          "cityId": {
            "type": "number",
            "example": 1
          },
          "image": {
            "type": "string",
            "example": "assets/images/home-cleaning/clean1.jpg"
          },
          "availabilitySlots": {
            "example": [
              "09:00 AM - 11:00 AM"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ownerProviderId": {
            "type": "string",
            "example": "PRO002"
          },
          "ownerProviderEmail": {
            "type": "string",
            "example": "fresh@example.com"
          }
        }
      },
...
      "VerificationDocument": {
        "type": "object",
        "properties": {
          "documentId": {
            "type": "string",
            "example": "DOC-PRO001-ID"
          },
          "documentType": {
            "type": "string",
            "example": "ID Proof"
          },
          "documentName": {
            "type": "string",
            "example": "Aadhaar Card"
          },
          "documentUrl": {
            "type": "string",
            "example": "/documents/providers/PRO001/aadhaar.pdf"
          },
          "documentStatus": {
            "type": "string",
            "example": "Approved",
            "enum": [
              "Pending",
              "Approved",
              "Rejected"
            ]
          },
          "required": {
            "type": "boolean",
            "example": true
          },
          "uploadedAt": {
            "type": "string",
            "example": "2026-03-08T10:30:00.000Z"
          },
          "rejectionReason": {
            "type": "string",
            "example": "Document is blurred."
          }
        },
        "required": [
          "documentId",
          "documentType",
          "documentName",
          "documentUrl",
          "documentStatus",
          "required",
          "uploadedAt"
        ]
      },
...
      "ProviderVerification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "PRO013"
          },
          "name": {
            "type": "string",
            "example": "Anita Verma"
          },
          "email": {
            "type": "string",
            "example": "anita.verma@email.com"
          },
          "organisationName": {
            "type": "string",
            "example": "Anita Beauty Studio"
          },
          "phone": {
            "type": "string",
            "example": "+91 9876543230"
          },
          "category": {
            "type": "string",
            "example": "Salon Services"
          },
          "experience": {
            "type": "number",
            "example": 5
          },
          "location": {
            "type": "string",
            "example": "Mumbai"
          },
          "address": {
            "type": "string",
            "example": "Andheri West, Mumbai"
          },
          "skills": {
            "example": [
              "Hair styling",
              "Facial care"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "certifications": {
            "example": [
              "Beauty academy certificate"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "completedJobs": {
            "type": "number",
            "example": 128
          },
          "rating": {
            "type": "number",
            "example": 4.7
          },
          "submittedDate": {
            "type": "string",
            "example": "2026-03-08T10:30:00.000Z"
          },
          "joinedDate": {
            "type": "string",
            "example": "2026-03-08T10:30:00.000Z"
          },
          "status": {
            "type": "string",
            "example": "Pending",
            "enum": [
              "Pending",
              "Verified",
              "Rejected",
              "Suspended"
            ]
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VerificationDocument"
            }
          },
          "statusHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProviderStatusHistoryItem"
            }
          },
          "rejectionReason": {
            "type": "string",
            "example": "Needs one clearer document."
          },
          "adminRemarks": {
            "type": "string",
            "example": "Called provider and verified service area."
          },
          "verifiedAt": {
            "type": "string",
            "example": "2026-03-10T10:30:00.000Z"
          },
          "suspendedAt": {
            "type": "string",
            "example": "2026-03-11T10:30:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "phone",
          "category",
          "experience",
          "location",
          "address",
          "skills",
          "certifications",
          "completedJobs",
          "rating",
          "submittedDate",
          "joinedDate",
          "status",
          "documents",
          "statusHistory"
        ]
      },
...
      "CreateProviderVerificationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "PRO018"
          },
          "name": {
            "type": "string",
            "example": "Fresh Clean Services"
          },
          "email": {
            "type": "string",
            "example": "fresh@example.com"
          },
          "organisationName": {
            "type": "string",
            "example": "Fresh Clean Services"
          },
          "phone": {
            "type": "string",
            "example": "9876501234"
          },
          "category": {
            "type": "string",
            "example": "Home cleaning"
          },
          "experience": {
            "type": "number",
            "example": 4
          },
          "location": {
            "type": "string",
            "example": "Chennai"
          },
          "address": {
            "type": "string",
            "example": "Anna Nagar, Chennai"
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateProviderDocumentDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "phone",
          "category",
          "experience",
          "location",
          "address",
          "documents"
        ]
      },
...
    }
  }
}